chatbot_service/requirements.txt
fastapi[all]
google-generativeai
sentence-transformers
faiss-cpu
numpy
